This version does not yet work, but here are the changes.
If there is a title, then the breadcrumbs display a title.
You have an optional viewname.  So you can
show the breadcrumbs for /index or for /manage

There is a much simpler introduction.  Just
do BasicBredcrumbs(item,request).render()

Of if that is not the right simple abstractin, the docs  breaks down the
way to split it up.

Anyhow you did a beatuful job on this repository, as you have done on the
rest of Cromlech.